Resolve 'frame is null' errors
As shown in pull request oyvindkinsey/easyXDM#120 and currently shown on
twitter.com home timeline when scrolling rapidly, the waitForReady
function can complain that the frame is null. The probable cause is that
the waitForReady method isn't torn down in the destroy method. So if an
iframe is destroyed before it initialises, it can throw this kind of
error.

twitter.com:
  `Uncaught TypeError: Cannot read property 'contentWindow' of null`

I've pulled out the function and added an unbind. I think this makes
sense.